refactor: proof entries in table
