+++ web/assets/js/react/modules/History/ProofHistory.tsx
@@ -9,45 +9,10 @@ import { computeVerificationDataCommitment } from "../../utils/aligned";
 import { shortenHash } from "../../utils/crypto";
 import { useProofSentMessageReader } from "../../hooks/useProofSentMessageReader";
 import { ProofEntryActionBtn } from "./ProofEntryActionBtn";
-
-type KeysForStatus = ProofSubmission["status"];
-
-const colorBasedOnStatus: {
-	[key in KeysForStatus]: string;
-} = {
-	submitted: "bg-accent-100/20 text-accent-100",
-	pending: "bg-yellow/20 text-yellow",
-	claimed: "bg-blue/20 text-blue",
-	failed: "bg-red/20 text-red",
-	underpriced: "bg-orange/20 text-orange",
-};
-
-const tooltipStyleBasedOnStatus: {
-	[key in KeysForStatus]: string;
-} = {
-	submitted: "bg-accent-100 text-black",
-	pending: "bg-yellow text-black",
-	claimed: "bg-blue text-white",
-	failed: "bg-red text-white",
-	underpriced: "bg-orange text-black",
-};
-
-const tooltipText: { [key in KeysForStatus]: string } = {
-	submitted: "Solution verified and ready to be submitted",
-	claimed: "Already submitted to leaderboard",
-	pending:
-		"You need to wait until its verified before submitting the solution",
-	failed: "The proof failed to be verified, you have to re-send it",
-	underpriced: "The proof is underpriced, we suggest you bump the fee",
-};
-
-const statusText: { [key in KeysForStatus]: string } = {
-	claimed: "Claimed",
-	submitted: "Ready",
-	pending: "Pending",
-	failed: "Failed",
-	underpriced: "Pending",
-};
+import {
+	ProofBatchMerkleRoot,
+	ProofStatusWithTooltipDesc,
+} from "../../components/Table/ProofBodyItems";
 
 type Props = {
 	leaderboard_address: Address;
@@ -92,7 +57,7 @@ export const ProofHistory = ({
 			// if the proof is pending and it has passed more than 6 hours
 			// mark it as underpriced
 			if (proof.status === "pending") {
-				if (timeAgoInHs(proof.status) > 6) {
+				if (timeAgoInHs(proof.insertedAt) > 6) {
 					proof.status = "underpriced";
 				}
 			}
@@ -100,44 +65,9 @@ export const ProofHistory = ({
 			return {
 				rows: [
 					<TableBodyItem text={proof.game} />,
-					<td>
-						<div
-							className={`relative group/tooltip flex flex-row gap-2 items-center rounded px-1 w-fit ${
-								colorBasedOnStatus[proof.status]
-							}`}
-						>
-							<span className="hero-information-circle solid size-5"></span>
-							<p>{statusText[proof.status]}</p>
-
-							<div
-								className={`${
-									tooltipStyleBasedOnStatus[proof.status]
-								} rounded absolute mt-2 rounded -left-1/2 top-full mb-2 text-sm rounded px-2 py-1 opacity-0 group-hover/tooltip:opacity-100 transition pointer-events-none`}
-								style={{ width: 300, zIndex: 10000 }}
-							>
-								<p className="text-center text-xs">
-									{tooltipText[proof.status]}
-								</p>
-							</div>
-						</div>
-					</td>,
+					<ProofStatusWithTooltipDesc proof={proof} />,
 					<TableBodyItem text={timeAgo(proof.insertedAt)} />,
-					<td>
-						{proof.batchData?.batch_merkle_root ? (
-							<a
-								href={`https://explorer.alignedlayer.com/batches/${proof.batchData.batch_merkle_root}`}
-								className="underline"
-							>
-								{shortenHash(
-									bytesToHex(
-										proof.batchData.batch_merkle_root
-									)
-								)}
-							</a>
-						) : (
-							<p>...</p>
-						)}
-					</td>,
+					<ProofBatchMerkleRoot proof={proof} />,
 					<TableBodyItem
 						text={shortenHash(
 							bytesToHex(
